moved logout request from navbar to user info
