+++ client/src/Components/NavBar/index.js
@@ -1,6 +1,5 @@
 import React, { useContext, useState } from 'react';
 import { useHistory } from 'react-router-dom';
-import Axios from 'axios';
 import { Layout, Typography, Menu } from 'antd';
 import AuthProvider from '../../Context/Authentication';
 import Public from './Public';
@@ -12,25 +11,12 @@ const { Header } = Layout;
 const { Title } = Typography;
 
 const NavBar = () => {
-  const { setIsAuth, isAuth } = useContext(AuthProvider);
+  const { isAuth } = useContext(AuthProvider);
   const history = useHistory();
-
-  const [isLoading, setLoading] = useState(false);
-  const [errMsg, setErrMsg] = useState('');
   const [current, setCurrent] = useState('home');
 
   const handleMenu = (e) => {
     setCurrent(e.key);
-  };
-
-  const handleClick = async () => {
-    try {
-      setLoading(true);
-      await Axios('api/v1/logout');
-      setIsAuth(false);
-    } catch (err) {
-      setErrMsg('set interval error');
-    }
   };
 
   <Header>
@@ -54,11 +40,10 @@ const NavBar = () => {
       ) : (
         <>
           <User />
-          <UserInfo handleLogout={handleClick} isLoading={isLoading} />
+          <UserInfo />
         </>
       )}
     </Menu>
-    {errMsg ? <p className="error-msg">{errMsg}</p> : null}
   </Header>;
 };
